fix: make generated AST parse roots relocatable

+++ ClangAstParser/src/pt/up/fe/specs/clang/codeparser/ParallelCodeParser.java
@@ -403,27 +403,19 @@ public class ParallelCodeParser extends CodeParser {
             return null;
         }
 
-        ClangAstData clangParserData = clangParser.parse(sourceFile, id, standard, options);
+        try {
+            ClangAstData clangParserData = clangParser.parse(sourceFile, id, standard, options);
 
-        if (get(SHOW_CLANG_DUMP)) {
-            // SpecsLogs.msgInfo("Clang Dump:\n" + SpecsIo.read(new File(ClangAstParser.getClangDumpFilename())));
-            // SpecsLogs.msgInfo(clangParser.getClangDump());
-            clangDump.add(clangParser.getClangDump());
-        }
-
-        if (get(CLEAN)) {
-            // if (clangParser.getLastWorkingFolder() == null) {
-            // workingFolders.add(clangParser.getLastWorkingFolder());
-            // }
-            if (clangParser.getLastWorkingFolder() == null) {
-                SpecsLogs.msgInfo("No working folder found for source file '" + sourceFile + "'");
-            } else {
-                SpecsIo.deleteFolder(clangParser.getLastWorkingFolder());
+            if (get(SHOW_CLANG_DUMP)) {
+                clangDump.add(clangParser.getClangDump());
             }
 
+            return clangParserData;
+        } finally {
+            if (get(CLEAN) && clangParser.getLastWorkingFolder() != null) {
+                SpecsIo.deleteFolder(clangParser.getLastWorkingFolder());
+            }
         }
-
-        return clangParserData;
     }
 
     /*
